fix some NPE

+++ riverock-portlet/src/java/org/riverock/portlet/member/currency/CurrencyStdCursClassQuery.java
@@ -122,7 +122,7 @@ public class CurrencyStdCursClassQuery extends BaseClassQuery
 
         if (cat.isDebugEnabled())
 
-            cat.debug("idCurrency - "+idCurrency.longValue());
+            cat.debug("idCurrency - "+idCurrency);
 
     }
